requirements.txt
pydantic
requests
pillow
loguru
aiohttp
urllib3
